feat
@PathVariable("id") 수정

+++ src/main/java/com/fls/animecommunity/animesanctuary/service/BoardService.java
@@ -3,6 +3,7 @@ package com.fls.animecommunity.animesanctuary.service;
 import com.fls.animecommunity.animesanctuary.model.board.Board;
 import com.fls.animecommunity.animesanctuary.model.board.dto.BoardRequestsDto;
 import com.fls.animecommunity.animesanctuary.model.board.dto.BoardResponseDto;
+import com.fls.animecommunity.animesanctuary.model.board.dto.SuccessResponseDto;
 import com.fls.animecommunity.animesanctuary.repository.BoardRepository;
 
 import lombok.NoArgsConstructor;
@@ -27,7 +28,8 @@ public class BoardService {
         return boardRepository.findAllByOrderByModifiedAtDesc().stream().map(BoardResponseDto::new).toList();
     }
     
-    //게시글의 id를 가진 데이터를 boardRepository에서 찾아서 BoardResponseDto 객체로 만들어 반환한다.
+    //find
+    @Transactional
     public BoardResponseDto getPost(Long id) {
     	return boardRepository.findById(id).map(BoardResponseDto::new).orElseThrow(
                 () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
@@ -44,15 +46,29 @@ public class BoardService {
     }
     
     //update
-    public Board updatePost(Long id, Board postDetails) {
-        Board post = boardRepository.findById(id).orElseThrow(() -> new RuntimeException("Post not found"));
-        post.setTitle(postDetails.getTitle());
-        post.setContents(postDetails.getContents());
-        return boardRepository.save(post);
+    @Transactional
+    public BoardResponseDto updatePost(Long id, BoardRequestsDto requestsDto) throws Exception {
+        Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+        );
+        if (!requestsDto.getPassword().equals(board.getPassword()))
+            throw new Exception("비밀번호가 일치하지 않습니다.");
+
+        board.update(requestsDto);
+        return new BoardResponseDto(board);
     }
     
     //delete
-    public void deletePost(Long id) {
-    	boardRepository.deleteById(id);
+    @Transactional
+    public SuccessResponseDto deletePost(Long id, BoardRequestsDto requestsDto) throws Exception{
+    	Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+        );
+
+        if (!requestsDto.getPassword().equals(board.getPassword()))
+            throw new Exception("비밀번호가 일치하지 않습니다.");
+
+        boardRepository.deleteById(id);
+        return new SuccessResponseDto(true);
     }
 }
